``` refactor(template): make identifier optional in StandardCms component
Change the `identifier` prop in the `StandardCms` component to be optional
to improve the flexibility of its usage. Additionally, replace `shallowRef`
with `ref` for the `list` state management variable in the `genCmsTemplateSingleData`
function to simplify the data handling. These changes aim to enhance the
consistency and usability of the component and its related functions.
```